particle background and file organization
