+++ src/components/ParticleBackground.js
@@ -2,5 +2,5 @@ import Particles from 'react-particles-js';
 import particlesConfig from '../config/particle-config';
 
 export default function ParticleBackground() {
-	return <Particles></Particles>;
+	return <Particles params={particlesConfig}></Particles>;
 }
