+++ src/components/pages/Messenger.js
@@ -1,0 +1,5 @@
+import '../../styles/messenger.css';
+
+export default function Messenger() {
+	return <div className="messenger">hello</div>;
+}
