+++ src/styles/register.css
@@ -1,84 +1,92 @@
 .register {
-    width: 100vw;
-    height: 100vh;
-    background-color: #f0f2f5;
-    display: flex;
-    align-items: center;
-    justify-content: center;
+	width: 100vw;
+	height: 100vh;
+	background-color: #f0f2f5;
+	display: flex;
+	align-items: center;
+	justify-content: center;
 }
 
 .registerWrapper {
-    width: 70%;
-    height: 70%;
-    display: flex;
+	width: 70%;
+	height: 70%;
+	display: flex;
+	z-index: 20;
 }
 
-.registerLeft, .registerRight {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
+.registerLeft,
+.registerRight {
+	flex: 1;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
 }
 
 .registerLogo {
-    font-size: 50px;
-    font-weight: 800;
-    color: rgb(0, 162, 255);
-    margin-bottom: 10px;
+	font-size: 50px;
+	font-weight: 800;
+	color: rgb(0, 162, 255);
+	margin-bottom: 10px;
 }
 
 .registerDesc {
-    font-size: 24px;
+	font-size: 24px;
 }
 
 .registerBox {
-    height: 400px;
-    padding: 20px;
-    background-color: white;
-    border-radius: 10px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
+	height: 400px;
+	padding: 20px;
+	background-color: white;
+	border-radius: 10px;
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
 }
 
 .registerInput {
-    height: 50px;
-    border-radius: 10px;
-    border: 1px solid gray;
-    font-size: 18px;
-    padding-left: 20px;
+	height: 50px;
+	border-radius: 10px;
+	border: 1px solid gray;
+	font-size: 18px;
+	padding-left: 20px;
 }
 
 .registerInput:focus {
-    outline: none;
+	outline: none;
 }
 
 .registerButton {
-    height: 50px;
-    border-radius: 10px;
-    border: none;
-    background-color: rgb(0, 162, 255);
-    color: white;
-    font-size: 20px;
-    font-weight: 500;
-    cursor: pointer;
+	height: 50px;
+	border-radius: 10px;
+	border: none;
+	background-color: rgb(0, 162, 255);
+	color: white;
+	font-size: 20px;
+	font-weight: 500;
+	cursor: pointer;
 }
 
 .registerForgot {
-    text-align: center;
-    color: rgb(0, 162, 255);
+	text-align: center;
+	color: rgb(0, 162, 255);
 }
 
 .registerRegisterButton {
-    width: 60%;
-    align-self: center;
-    height: 50px;
-    border-radius: 10px;
-    border: none;
-    background-color: #42b72a;
-    color: white;
-    font-size: 20px;
-    font-weight: 500;
-    cursor: pointer;
+	width: 60%;
+	align-self: center;
+	height: 50px;
+	border-radius: 10px;
+	border: none;
+	background-color: #42b72a;
+	color: white;
+	font-size: 20px;
+	font-weight: 500;
+	cursor: pointer;
+}
 
+.particles {
+	position: absolute;
+	width: 100vw;
+	height: 100vh;
+	z-index: 0;
 }
